feat(dapp_register): move metamask wallet to separate component
