+++ app/packages/react-app/src/components/MetamaskConnect.tsx
@@ -1,0 +1,48 @@
+import { shortenAddress, useEthers, useLookupAddress } from "@usedapp/core";
+import React, { useEffect, useState } from "react";
+import Box from "@mui/material/Box";
+import logo from "../ethereumLogo.png";
+import Button from "@mui/material/Button";
+import { Header, Body, Image, Link } from "../components";
+import Editor from "../components/Editor";
+import { Container } from "@mui/material";
+
+function MetamaskConnect() {
+  const [rendered, setRendered] = useState("");
+
+  const ens = useLookupAddress();
+  const { account, activateBrowserWallet, deactivate, error } = useEthers();
+
+  useEffect(() => {
+    if (ens) {
+      setRendered(ens);
+    } else if (account) {
+      setRendered(shortenAddress(account));
+    } else {
+      setRendered("");
+    }
+  }, [account, ens, setRendered]);
+
+  useEffect(() => {
+    if (error) {
+      console.error("Error while connecting wallet:", error.message);
+    }
+  }, [error]);
+
+  return (
+    <Button
+      onClick={() => {
+        if (!account) {
+          activateBrowserWallet();
+        } else {
+          deactivate();
+        }
+      }}
+    >
+      {rendered === "" && "Connect Wallet"}
+      {rendered !== "" && rendered}
+    </Button>
+  );
+}
+
+export default MetamaskConnect;
